models: use the jsonToGraphQLQuery lib
Allows using real JS objects to generate GraphQL queries. Ultimate
reason to decide to move to this was that it was becoming to difficult
to manually serialise big JSON object for the Scrape table's JSONB
column. But also in general using real objects to build the queries is
just good all round.