web: Persist login state
Note that this abuses the already controversial JWT standards. Here we
are storing the token in localStorage. This is not recommended. Rather a
HTTP-only cookie containing a 'refresh token' should be stored. We will
need to update our JWT server to do that. For now we don't have
significant security requirements, hence crudely storing the master
token in the browser is acceptable.