[crawlers] fix big lat/lon
it was 2 separate lat/lon ordering bugs working in unison. the first
bug put 75% of our restaurants in the antarctic, the second bug was
that the yelp scraper also had the wrong ordering meaning that it
actually "correctly" placed restaurants in sanfran. it's only because
of that yelp bug that i never noticed the first bug. meaning that the
restaurants we see now are only 25% of our total restaurants, they
being the ones from yelp's bug.