[db] refactor to restaurant_taxonomy junction
This is the recommended way for many-to-many relationships to be
architected in Hasura. This maybe the most elegant solution but it
doesn't give us any ability to optimise the performance of critical
taxontaxonomy searches. I suspect the ideal solution is just to move to
a dedicated search service if performance becomes an issue.