[crawlers] the "Nopa" case
This was the restaurant that prompted me to find a bug in DB data,
there's no fix in the code, but there was a lot of research finding the
right SQL queries to clean up the DB.

Basically it seems there was scrape data from weeks and months ago that
conflicted with more recent scrape data during the internal merge stage
of our crawlers. In the case of "Nopa" there was a scrape that had the
correct name and address but the coordinates were in Yukaton! This meant
that when the correct source data tried to merge there'd be a constraint
error on the restaurant table (restaurant_name_address_key). My solution
was to delete all restaurants that didn't have any associated scrape
data - these were from a time when we had a different method of
associating scrape records with restaurants. And to also delete
restaurants that only had 1 associated scrape record and were created a
month or more ago.

An example of one of the SQL queries is here:

delete from restaurant where id in
  (select id from
    (
      select
        r.id as id,
        left(r.name, 20) as name,
        count(s.restaurant_id),
        date_trunc('day', max(s.created_at)) as crawl_day,
        r.created_at
      from restaurant r
      left join scrape s on s.restaurant_id = r.id
      group by r.id
      order by crawl_day desc nulls last
    ) as rs
  where count = 0
  );