[crawlers] yelp: reduce the size of MAPVIEW
This is a probably equivalent to "zooming in". Potentially this will
make the crawler find more restaurants. Maybe, it's just an experiment.
Currently the crawler is missing some restaurants that seem to be easy
to crawl when manually focussing on a single geo coord.

This does not solve the problem of the Yelp geo search endpoint
sometimes just not finding the restaurant at all, even when manually
focussing on a single coord.