[graph] First fix for JSONB support
Interestingly GQLess is ok about reading JSONB from the DB but not about
inserting or updating it. So this commit includes a patch to GQLess to
prevent JSON fields being converted into serialised JSON strings,
instead merely stripping JSON keys of their double quotes.

 #yodawg I heard you like patches in yo patches, so I patched the patch
 to make a new patch.

This also includes 2 tweaks to resolveFields() behaviour. Firstly it's
updated to deal with results as arrays. And secondly the
touchToResolveInGQLess() part is refactored into its own function,
because not only do the fields need to be touched in order to build the
returning{} part of the query, but they also need to be touched once the
result is returned.

@natew I don't like the @ts-ignore in touchToResolveInGQLess(), I'm sure
there must be a better way.