[crawlers] Yelp search filter bug fix
in all our collect() code there's a function isMutableField(), the bug
i kept seeing was that a null field would get passed to it, but this
would only happen on production, so difficult to debug. i'm pretty sure
now though that the reason this happens is because the production code
doesn't have the updated schemas from a yarn generate. i've gotten into
a habit of hot deploying migrations to hasura, it's really quick and
easy. but what this hot migration doesn't do is include the associated
gqless schema updates. so there's a discrepancy between the object
fields that the crawler generates for persistence and the gqless schema
fields embedded in the compiled code (that require a proper docker
build to deploy). so lesson learned. i think that could have caused
a huge number of internal crawls to fail over the last few months