[crawlers] Prevent flakey Yelp failure
There are some long-running parts of a yelp crawl such as going through
the pages of photos and reviews. These should bail early during tests,
as we don't need to go through them all for the tests to be effective.
Otherwise the pagination can take so long as to trigger the test timeout