[crawlers] Save geocoder_id at restaurant creation
There already existed a logic of adding the geocoder_ids in bulk through
a worker job. But after being away from the code for 6 months I was suddenly
struck by the oddity of that approach. It seems much more natural to be
persisting the geocoder_id at the point when a canonical restaurant is first
created. There may be some non-obvious gotcha to this seemingly more intuitive
approach, so for now I will leave the `updateAllGeocoderIDs`` code.