[crawlers] Detect restaurant source ID change
This is potentially a critical part of maintaining internal consistency
of canonical restaurant definitions. Currently there is no attempt to
100% guarantee that scrape data always gets applied to the same
restaurant. Ultimately we either need to attempt that or consistently
handle source changes. For now, this is a first step, we just make a
note when a source ID changes